Refactor jupyter_tikz module into focused components with facade compatibility
